Put the log and ddtrace info on a single line [TER-11370]
With the previous change, it wasn't obvious that the `lograge_format`
variable included a newline. So when I added additional log data after,
it showed up on a new line. This was not what we want. The options were
to just place the dd info first rather than last, or build it into the
data, rather than using the supplied variable from datadog. I just went
with putting it first.